Fix a bug in the prev/next half step SE generation from CE
When the CE is on the odd plane, the associated SEs at the previous and next half step plane should use one more index.  This bug caused the plane wave to move in the wrong direction.

After fixing the bug, the periodic boundary condition treatment is successfully added to the Linear Scalar solver.  A plane-wave test case is added to the unit test suite.