Allow Python-based callbacks in the basic Selm and Solver
The following Selm member functions are allowed to be changed by assigning Python functions to Solver object: xn(), xp(), tn(), tp(), so0p(), and update_cfl().

To void increased footprint of Selm (and its derived classes), the Python-customized functions are added to the Field object, which is directly accessible by Selm class.  The Python interface to access the Python-customized functions are made on the Solver class.

Other changes:

1. Fix a random failiure in tests/test_celm_selm.py::CelmTC::test_calc_so.  It seems to be caused by the lack of field initialization.
2. Make zero test more accurate for ce.calc_so using 14 decimal points